api_4 fix itemsのリファクタ及びbrewery関連の実装
