+++ app/Services/ItemService.php
@@ -1,0 +1,153 @@
+<?php
+
+namespace App\Services;
+
+use App\Http\Resources\ShowItemsResource;
+use App\Http\Resources\ArrayDataResource;
+use App\Models\Item;
+use App\Models\Brewery;
+use App\Models\Country;
+use App\Models\Prefecture;
+use App\Models\Taste;
+use App\Models\Container;
+use App\Models\Style;
+use App\Models\Color;
+use App\Models\Abv;
+use App\Models\Type;
+
+class ItemService
+{
+    public function searchItems($request)
+    {
+        // アイテムを抽出するクエリビルダを作成
+        $query = Item::query();        
+        // 検索条件取得
+        $fetchKeywordValue = $request->input('inputKeyword');
+
+        // inputKeyword が渡されたらキーワード検索、そうでなければ条件検索を実行
+        if (!empty($fetchKeywordValue)) {
+            // キーワード検索
+            // 検索文字列全体の前後にある空白を除去
+            $keyword_remove_space = preg_replace( '/\A[\p{C}\p{Z}]++|[\p{C}\p{Z}]++\z/u', '', $fetchKeywordValue);
+            // 検索文字列内の半角スペースを全角スペースにする
+            $keyword_unify_space =  mb_convert_kana($keyword_remove_space, 's');
+            // 全角空白で文字を区切り配列へ
+            $searchTerms = preg_split('/[\s]+/', $keyword_unify_space);
+
+            foreach ($searchTerms as $searchTerm) {
+                $query->where('name', 'like', "%{$searchTerm}%");
+                $query->orWhereHas('brewery', function ($query) use ($searchTerm) {
+                    $query->where('name', 'like', "%$searchTerm%");
+                });
+                $query->orWhereHas('country', function ($query) use ($searchTerm) {
+                    $query->where('name', 'like', "%$searchTerm%");
+                });
+                $query->orWhereHas('prefecture', function ($query) use ($searchTerm) {
+                    $query->where('name', 'like', "%$searchTerm%");
+                });
+                $query->orWhereHas('style', function ($query) use ($searchTerm) {
+                    $query->where('name', 'like', "%$searchTerm%");
+                });
+                $query->orWhereHas('color', function ($query) use ($searchTerm) {
+                    $query->where('name', 'like', "%$searchTerm%");
+                });
+                $query->orWhereHas('abv', function ($query) use ($searchTerm) {
+                    $query->where('name', 'like', "%$searchTerm%");
+                });
+                $query->orWhereHas('type', function ($query) use ($searchTerm) {
+                    $query->where('name', 'like', "%$searchTerm%");
+                });
+                $query->orwhereHas('tastes', function ($subQuery) use ($searchTerm) {
+                    $subQuery->where('name', 'like', "%$searchTerm%");
+                });
+                $query->orwhereHas('containers', function ($subQuery) use ($searchTerm) {
+                    $subQuery->where('name', 'like', "%$searchTerm%");
+                });
+            }
+
+        } else {
+            // 条件検索
+            $fetchBreweryId = $request->input('breweryId');
+            $fetchCountryId = $request->input('countryId');
+            $fetchPrefectureId = $request->input('prefectureId');
+            $fetchTasteId = $request->input('tasteId');
+            $fetchContainerId = $request->input('containerId');
+            $fetchStyleId = $request->input('styleId');
+            $fetchColorId = $request->input('colorId');
+            $fetchAbvId = $request->input('abvId');
+            $fetchTypeId = $request->input('typeId');
+
+            // 各条件を組み合わせてクエリを絞り込む
+            if (!empty($fetchBreweryId)) {
+                $query->whereIn('brewery_id', $fetchBreweryId);
+            }
+            if (!empty($fetchCountryId)) {
+                $query->whereIn('country_id', $fetchCountryId);
+            }
+            if (!empty($fetchPrefectureId)) {
+                $query->whereIn('prefecture_id', $fetchPrefectureId);
+            }
+            if (!empty($fetchStyleId)) {
+                $query->whereIn('style_id', $fetchStyleId);
+            }
+            if (!empty($fetchColorId)) {
+                $query->whereIn('color_id', $fetchColorId);
+            }
+            if (!empty($fetchAbvId)) {
+                $query->whereIn('abv_id', $fetchAbvId);
+            }
+            if (!empty($fetchTypeId)) {
+                $query->whereIn('type_id', $fetchTypeId);
+            }
+
+            // アイテムを抽出するクエリビルダを作成
+            // tastesテーブルの条件を絞り込む
+            if (!empty($fetchTasteId)) {
+                $query->whereHas('tastes', function ($subQuery) use ($fetchTasteId) {
+                    $subQuery->whereIn('tastes.id', $fetchTasteId);
+                });
+            }
+            // containersテーブルの条件を絞り込む
+            if (!empty($fetchContainerId)) {
+                $query->whereHas('containers', function ($subQuery) use ($fetchContainerId) {
+                    $subQuery->whereIn('containers.id', $fetchContainerId);
+                });
+            }
+        }
+
+        // クエリの結果を取得
+        $itemIds = $query->pluck('id');
+        $items = ShowItemsResource::collection(Item::find($itemIds));
+
+        return $items;
+    }
+
+    public function getInitialData()
+    {
+        // 初期選択データを各テーブルから全取得する
+        $brewery = Brewery::all();
+        $country = Country::all();
+        $prefecture = Prefecture::all();
+        $taste = Taste::all();
+        $container = Container::all();
+        $style = Style::all();
+        $color = Color::all();
+        $abv = Abv::all();
+        $type = Type::all();
+
+        $initialData = [
+            'brewery' => ArrayDataResource::collection($brewery),
+            'country' => ArrayDataResource::collection($country),
+            'prefecture' => ArrayDataResource::collection($prefecture),
+            'taste' => ArrayDataResource::collection($taste),
+            'container' => ArrayDataResource::collection($container),
+            'style' => ArrayDataResource::collection($style),
+            'color' => ArrayDataResource::collection($color),
+            'abv' => ArrayDataResource::collection($abv),
+            'type' => ArrayDataResource::collection($type),
+        ];
+
+        return $initialData;
+    }
+    
+}
