+++ routes/api.php
@@ -2,6 +2,9 @@
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\ItemsController;
+use App\Http\Controllers\BreweriesController;
+use App\Http\Controllers\ItemConditionsController;
 
 /*
 |--------------------------------------------------------------------------
@@ -18,15 +21,11 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-// items
-// Route::resource('items', 'ItemsController');
-Route::get('items/create', [App\Http\Controllers\ItemsController::class, 'create']);
-Route::post('items', [App\Http\Controllers\ItemsController::class, 'store']);
-Route::post('items/search', [App\Http\Controllers\ItemsController::class, 'search']);
-Route::get('items/{id}', [App\Http\Controllers\ItemsController::class, 'show']);
+Route::resources([
+    'items' => ItemsController::class,
+    'breweries' => BreweriesController::class,
+]);
 
-// breweries
-Route::get('breweries/create', [App\Http\Controllers\BreweriesController::class, 'create']);
-Route::post('breweries', [App\Http\Controllers\BreweriesController::class, 'store']);
-Route::post('breweries/search', [App\Http\Controllers\BreweriesController::class, 'search']);
-Route::get('breweries/{id}', [App\Http\Controllers\BreweriesController::class, 'show']);
+// Item Conditions
+Route::get('items/conditions/index', [ItemConditionsController::class, 'index']);
+Route::post('items/conditions/result', [ItemConditionsController::class, 'result']);
